+++ CHANGES.rst
@@ -18,6 +18,7 @@ The ASDF Standard is at v1.6.0
 - Remove deprecated TagDefinition.schema_uri [#1595]
 - Removed deprecated AsdfFile.open and deprecated asdf.open
   AsdfFile.write_to and AsdfFile.update kwargs [#1592]
+- Fix ``AsdfFile.info`` loading all array data [#1572]
 
 2.15.1 (2023-08-07)
 -------------------